Refactor, move common parts to packages
