--- pkg/keycloak/proxy/stores.go
@@ -1,42 +1,0 @@
-/*
-Copyright 2015 All rights reserved.
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package proxy
-
-import (
-	"context"
-
-	"github.com/gogatekeeper/gatekeeper/pkg/apperrors"
-	"github.com/gogatekeeper/gatekeeper/pkg/storage"
-	"github.com/gogatekeeper/gatekeeper/pkg/utils"
-)
-
-// Get retrieves a token from the store, the key we are using here is the access token
-func GetRefreshTokenFromStore(
-	ctx context.Context,
-	store storage.Storage,
-	token string,
-) (string, error) {
-	// step: the key is the access token
-	val, err := store.Get(ctx, utils.GetHashKey(token))
-	if err != nil {
-		return val, err
-	}
-	if val == "" {
-		return val, apperrors.ErrNoSessionStateFound
-	}
-
-	return val, nil
-}
