+++ pkg/utils/token.go
@@ -1,0 +1,252 @@
+package utils
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"net/http"
+	"regexp"
+	"strings"
+	"time"
+
+	oidc3 "github.com/coreos/go-oidc/v3/oidc"
+	"github.com/go-jose/go-jose/v3/jwt"
+	"github.com/gogatekeeper/gatekeeper/pkg/apperrors"
+	"github.com/gogatekeeper/gatekeeper/pkg/proxy/metrics"
+	"github.com/gogatekeeper/gatekeeper/pkg/proxy/models"
+	"go.uber.org/zap"
+	"go.uber.org/zap/zapcore"
+	"golang.org/x/oauth2"
+)
+
+func VerifyToken(
+	ctx context.Context,
+	provider *oidc3.Provider,
+	rawToken string,
+	clientID string,
+	skipClientIDCheck bool,
+	skipIssuerCheck bool,
+) (*oidc3.IDToken, error) {
+	// This verifier with this configuration checks only signatures
+	// we want to know if we are using valid token
+	// bad is that Verify method doesn't check first signatures, so
+	// we have to do it like this
+	verifier := provider.Verifier(
+		&oidc3.Config{
+			ClientID:          clientID,
+			SkipClientIDCheck: true,
+			SkipIssuerCheck:   true,
+			SkipExpiryCheck:   true,
+		},
+	)
+	_, err := verifier.Verify(ctx, rawToken)
+	if err != nil {
+		return nil, errors.Join(apperrors.ErrTokenSignature, err)
+	}
+
+	// Now doing expiration check
+	verifier = provider.Verifier(
+		&oidc3.Config{
+			ClientID:          clientID,
+			SkipClientIDCheck: skipClientIDCheck,
+			SkipIssuerCheck:   skipIssuerCheck,
+			SkipExpiryCheck:   false,
+		},
+	)
+
+	oToken, err := verifier.Verify(ctx, rawToken)
+	if err != nil {
+		return nil, err
+	}
+
+	return oToken, nil
+}
+
+func ParseRefreshToken(rawRefreshToken string) (*jwt.Claims, error) {
+	refreshToken, err := jwt.ParseSigned(rawRefreshToken)
+	if err != nil {
+		return nil, err
+	}
+
+	stdRefreshClaims := &jwt.Claims{}
+	err = refreshToken.UnsafeClaimsWithoutVerification(stdRefreshClaims)
+	if err != nil {
+		return nil, err
+	}
+
+	return stdRefreshClaims, nil
+}
+
+// GetRefreshedToken attempts to refresh the access token, returning the parsed token, optionally with a renewed
+// refresh token and the time the access and refresh tokens expire
+//
+// NOTE: we may be able to extract the specific (non-standard) claim refresh_expires_in and refresh_expires
+// from response.RawBody.
+// When not available, keycloak provides us with the same (for now) expiry value for ID token.
+func GetRefreshedToken(
+	ctx context.Context,
+	conf *oauth2.Config,
+	httpClient *http.Client,
+	oldRefreshToken string,
+) (jwt.JSONWebToken, string, string, time.Time, time.Duration, error) {
+	ctx = context.WithValue(ctx, oauth2.HTTPClient, httpClient)
+	start := time.Now()
+
+	tkn, err := conf.TokenSource(ctx, &oauth2.Token{RefreshToken: oldRefreshToken}).Token()
+	if err != nil {
+		if strings.Contains(err.Error(), "invalid_grant") {
+			return jwt.JSONWebToken{},
+				"",
+				"",
+				time.Time{},
+				time.Duration(0),
+				apperrors.ErrRefreshTokenExpired
+		}
+		return jwt.JSONWebToken{},
+			"",
+			"",
+			time.Time{},
+			time.Duration(0),
+			err
+	}
+
+	taken := time.Since(start).Seconds()
+	metrics.OauthTokensMetric.WithLabelValues("renew").Inc()
+	metrics.OauthLatencyMetric.WithLabelValues("renew").Observe(taken)
+
+	token, err := jwt.ParseSigned(tkn.AccessToken)
+	if err != nil {
+		return jwt.JSONWebToken{},
+			"",
+			"",
+			time.Time{},
+			time.Duration(0),
+			err
+	}
+
+	refreshToken, err := jwt.ParseSigned(tkn.RefreshToken)
+	if err != nil {
+		return jwt.JSONWebToken{},
+			"",
+			"",
+			time.Time{},
+			time.Duration(0),
+			err
+	}
+
+	stdClaims := &jwt.Claims{}
+	err = token.UnsafeClaimsWithoutVerification(stdClaims)
+	if err != nil {
+		return jwt.JSONWebToken{},
+			"",
+			"",
+			time.Time{},
+			time.Duration(0),
+			err
+	}
+
+	refreshStdClaims := &jwt.Claims{}
+	err = refreshToken.UnsafeClaimsWithoutVerification(refreshStdClaims)
+	if err != nil {
+		return jwt.JSONWebToken{},
+			"",
+			"",
+			time.Time{},
+			time.Duration(0),
+			err
+	}
+
+	refreshExpiresIn := time.Until(refreshStdClaims.Expiry.Time())
+
+	return *token,
+		tkn.AccessToken,
+		tkn.RefreshToken,
+		stdClaims.Expiry.Time(),
+		refreshExpiresIn,
+		nil
+}
+
+// CheckClaim checks whether claim in userContext matches claimName, match. It can be String or Strings claim.
+//
+//nolint:cyclop
+func CheckClaim(
+	logger *zap.Logger,
+	user *models.UserContext,
+	claimName string,
+	match *regexp.Regexp,
+	resourceURL string,
+) bool {
+	errFields := []zapcore.Field{
+		zap.String("claim", claimName),
+		zap.String("access", "denied"),
+		zap.String("email", user.Email),
+		zap.String("resource", resourceURL),
+	}
+
+	lLog := logger.With(errFields...)
+	if _, found := user.Claims[claimName]; !found {
+		lLog.Warn("the token does not have the claim")
+		return false
+	}
+
+	switch user.Claims[claimName].(type) {
+	case []interface{}:
+		claims, assertOk := user.Claims[claimName].([]interface{})
+		if !assertOk {
+			logger.Error(apperrors.ErrAssertionFailed.Error())
+			return false
+		}
+
+		for _, v := range claims {
+			value, ok := v.(string)
+			if !ok {
+				lLog.Warn(
+					"Problem while asserting claim",
+					zap.String(
+						"issued",
+						fmt.Sprintf("%v", user.Claims[claimName]),
+					),
+					zap.String("required", match.String()),
+				)
+
+				return false
+			}
+
+			if match.MatchString(value) {
+				return true
+			}
+		}
+
+		lLog.Warn(
+			"claim requirement does not match any element claim group in token",
+			zap.String("issued", fmt.Sprintf("%v", user.Claims[claimName])),
+			zap.String("required", match.String()),
+		)
+
+		return false
+	case string:
+		claims, assertOk := user.Claims[claimName].(string)
+		if !assertOk {
+			logger.Error(apperrors.ErrAssertionFailed.Error())
+			return false
+		}
+		if match.MatchString(claims) {
+			return true
+		}
+
+		lLog.Warn(
+			"claim requirement does not match claim in token",
+			zap.String("issued", claims),
+			zap.String("required", match.String()),
+		)
+
+		return false
+	default:
+		logger.Error(
+			"unable to extract the claim from token not string or array of strings",
+		)
+	}
+
+	lLog.Warn("unexpected error")
+	return false
+}
